requirements.txt
requests>=2.28.0
python-whois>=0.8.0
dnspython>=2.2.0
geoip2>=4.6.0
ipaddress>=1.0.23 